+++ packages/x-components/src/x-installer/api/__tests__/default-api.spec.ts
@@ -1,7 +1,7 @@
 import { createLocalVue } from '@vue/test-utils';
 import { XComponentsAdapterDummy } from '../../../__tests__/adapter.dummy';
 import { XInstaller } from '../../x-installer/x-installer';
-import { SnippetConfig } from '../api.types';
+import { NormalisedSnippetConfig, SnippetConfig } from '../api.types';
 import { BaseXAPI } from '../base-api';
 import { XDummyBus } from '../../../__tests__/bus.dummy';
 
@@ -103,5 +103,19 @@ describe('testing default X API', () => {
     api?.setSnippetConfig({ store: 'Portugal' });
     await vue.nextTick();
     expect(storeElement).toHaveTextContent('Portugal');
+  });
+
+  it('should allow set the snippetConfig getter', () => {
+    const snippetConfig: NormalisedSnippetConfig = {
+      instance: 'test',
+      scope: 'test',
+      lang: 'es',
+      uiLang: 'es'
+    };
+    defaultXAPI?.setSnippetConfigGetter(() => snippetConfig);
+
+    const snippet = defaultXAPI?.getSnippetConfig();
+
+    expect(snippet).toEqual(snippetConfig);
   });
 });