+++ packages/x-components/src/x-installer/x-installer/x-installer.ts
@@ -126,6 +126,7 @@ export class XInstaller {
       this.api = api ?? new BaseXAPI();
       this.api.setInitCallback(this.init.bind(this));
       this.api.setSnippetConfigCallback(this.updateSnippetConfig.bind(this));
+      this.api.setSnippetConfigGetter(this.getSnippetConfig.bind(this));
       window.InterfaceX = this.api;
     }
   }
@@ -354,5 +355,16 @@ export class XInstaller {
     forEach(this.normaliseSnippetConfig(newSnippetConfig), (name, value) => {
       this.getVue().set(this.snippetConfig!, name, value);
     });
+  }
+
+  /**
+   * Getter for the snippet config object.
+   *
+   * @returns The {@link NormalisedSnippetConfig | snippetConfig} object.
+   *
+   * @public
+   */
+  protected getSnippetConfig(): NormalisedSnippetConfig {
+    return this.snippetConfig!;
   }
 }