+++ packages/x-components/src/x-modules/queries-preview/store/actions/fetch-and-save-query-preview.action.ts
@@ -14,15 +14,18 @@ import { QueriesPreviewXStoreModule } from '../types';
 export const fetchAndSaveQueryPreview: QueriesPreviewXStoreModule['actions']['fetchAndSaveQueryPreview'] =
   ({ dispatch, commit }, request) => {
     const { query } = request;
+
     if (!query) {
       return;
     }
+
     commit('setQueryPreview', {
       request,
       results: [],
       status: 'loading',
       totalResults: 0
     });
+
     return dispatch('fetchQueryPreview', request)
       .then(response => {
         commit('setQueryPreview', {