+++ packages/x-components/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.0-alpha.4](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.3...@empathyco/x-components@4.1.0-alpha.4) (2023-12-29)
+
+
+### Features
+
+* **x-plugin:** add snippet config getter to XAPI (#1378) ([a76e1b7](https://github.com/empathyco/x/commit/a76e1b716b9aa86fc6853bcf4d1bb7aaa9478f4a))
+
+
+
+## [4.1.0-alpha.3](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.2...@empathyco/x-components@4.1.0-alpha.3) (2023-12-29)
+
+
+### Documentation
+
+* optimize images for SEO (#1379) ([b115a08](https://github.com/empathyco/x/commit/b115a08caa450447f0515532c3d9c9fd1125a1a4))
+
+
+
+## [4.1.0-alpha.2](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.1...@empathyco/x-components@4.1.0-alpha.2) (2023-12-28)
+
+**Note:** Version bump only for package @empathyco/x-components
+
+
+
+
+
+## [4.1.0-alpha.1](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.0...@empathyco/x-components@4.1.0-alpha.1) (2023-12-27)
+
+**Note:** Version bump only for package @empathyco/x-components
+
+
+
+
+
+## [4.1.0-alpha.0](https://github.com/empathyco/x/compare/@empathyco/x-components@4.0.2-alpha.0...@empathyco/x-components@4.1.0-alpha.0) (2023-12-11)
+
+
+### Features
+
+* **query-preview:** use cached query-previews (#1360) ([2579ca4](https://github.com/empathyco/x/commit/2579ca42d52c6f6c428b9129d63dc4a0ddb4788c))
+
+
+
 ## [4.0.2-alpha.0](https://github.com/empathyco/x/compare/@empathyco/x-components@4.0.1...@empathyco/x-components@4.0.2-alpha.0) (2023-11-29)
 
 **Note:** Version bump only for package @empathyco/x-components